requirements.txt
azure-ai-documentintelligence
python-dotenv
flask
flask-cors
requests
